fix type/calc_tax

+++ src/type/calc_tax.py
@@ -1,2 +1,2 @@
 def calc_tax(price, per):
-    return price * per
+    return int(float(price) * float(per))
